+++ src/js/core/height-match.js
@@ -16,13 +16,12 @@ export default {
     },
 
     computed: {
-        elements({ target }, $el) {
-            return $$(target, $el);
-        },
+        elements: ({ target }, $el) => $$(target, $el),
     },
 
     observe: resize({
-        target: ({ $el, elements }) => [$el, ...elements],
+        target: ({ $el, elements }) =>
+            elements.reduce((elements, el) => elements.concat(el, ...el.children), [$el]),
     }),
 
     update: {